# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AERISYN – AI Weather Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CurrentWeather {
  temperature: number;
  feels_like: number;
  humidity: number;
  pressure: number;
  wind_speed: number;
  wind_direction: number;
  /** @nullable */
  wind_gusts?: number | null;
  visibility: number;
  cloud_cover: number;
  rainfall: number;
  /** @nullable */
  snowfall?: number | null;
  uv_index: number;
  weather_code: number;
  weather_description: string;
  is_day: number;
  sunrise: string;
  sunset: string;
  /** @nullable */
  moonrise?: string | null;
  /** @nullable */
  moonset?: string | null;
}

export interface HourlyWeather {
  time: string;
  temperature: number;
  /** @nullable */
  feels_like?: number | null;
  humidity: number;
  precipitation_probability: number;
  /** @nullable */
  precipitation?: number | null;
  weather_code: number;
  weather_description?: string;
  wind_speed: number;
  /** @nullable */
  pressure?: number | null;
  /** @nullable */
  uv_index?: number | null;
  is_day: number;
}

export interface DailyWeather {
  date: string;
  temp_max: number;
  temp_min: number;
  weather_code: number;
  weather_description?: string;
  sunrise: string;
  sunset: string;
  /** @nullable */
  uv_index_max?: number | null;
  precipitation_sum: number;
  precipitation_probability_max: number;
  wind_speed_max: number;
  /** @nullable */
  wind_gusts_max?: number | null;
}

export interface ForecastData {
  hourly: HourlyWeather[];
  daily: DailyWeather[];
}

export interface HistoricalWeather {
  daily: DailyWeather[];
  hourly: HourlyWeather[];
}

export type WeatherAlertSeverity = typeof WeatherAlertSeverity[keyof typeof WeatherAlertSeverity];


export const WeatherAlertSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  extreme: 'extreme',
} as const;

export interface WeatherAlert {
  id: string;
  type: string;
  severity: WeatherAlertSeverity;
  title: string;
  description: string;
  /** @nullable */
  start_time?: string | null;
  /** @nullable */
  end_time?: string | null;
}

export interface AirQualityHourly {
  time: string;
  aqi: number;
  pm2_5: number;
  pm10: number;
  /** @nullable */
  ozone?: number | null;
  /** @nullable */
  nitrogen_dioxide?: number | null;
}

export interface AirQuality {
  aqi: number;
  aqi_category?: string;
  pm2_5: number;
  pm10: number;
  carbon_monoxide: number;
  nitrogen_dioxide: number;
  ozone: number;
  sulphur_dioxide: number;
  /** @nullable */
  dust: number | null;
  uv_index: number;
  hourly?: AirQualityHourly[];
}

export interface GeocodingResult {
  /** @nullable */
  id?: number | null;
  name: string;
  lat: number;
  lon: number;
  country: string;
  /** @nullable */
  country_code?: string | null;
  /** @nullable */
  admin1?: string | null;
  /** @nullable */
  admin2?: string | null;
  /** @nullable */
  timezone?: string | null;
  /** @nullable */
  elevation?: number | null;
  /** @nullable */
  population?: number | null;
}

export interface FavoriteCity {
  id: number;
  name: string;
  lat: number;
  lon: number;
  country: string;
  /** @nullable */
  country_code?: string | null;
  /** @nullable */
  admin1?: string | null;
  /** @nullable */
  timezone?: string | null;
  added_at: string;
}

export interface FavoriteCityInput {
  name: string;
  lat: number;
  lon: number;
  country: string;
  /** @nullable */
  country_code?: string | null;
  /** @nullable */
  admin1?: string | null;
  /** @nullable */
  timezone?: string | null;
}

export interface SearchHistoryItem {
  id: number;
  name: string;
  lat: number;
  lon: number;
  country: string;
  /** @nullable */
  country_code?: string | null;
  /** @nullable */
  admin1?: string | null;
  /** @nullable */
  timezone?: string | null;
  searched_at: string;
}

export interface SearchHistoryInput {
  name: string;
  lat: number;
  lon: number;
  country: string;
  /** @nullable */
  country_code?: string | null;
  /** @nullable */
  admin1?: string | null;
  /** @nullable */
  timezone?: string | null;
}

export interface MlPrediction {
  target: string;
  predicted_value: number;
  unit: string;
  confidence: number;
  model_used: string;
  /** @nullable */
  lower_bound?: number | null;
  /** @nullable */
  upper_bound?: number | null;
}

export interface MlPredictions {
  location_lat: number;
  location_lon: number;
  predictions: MlPrediction[];
  generated_at: string;
}

export interface MlModelMetric {
  model_name: string;
  rmse: number;
  mae: number;
  r2: number;
  accuracy: number;
  /** @nullable */
  training_samples?: number | null;
  /** @nullable */
  last_trained?: string | null;
}

export interface MlModelComparison {
  models: MlModelMetric[];
  best_model: string;
  target_variable?: string;
}

export interface MlTrainInput {
  lat: number;
  lon: number;
  days_of_history?: number;
}

export interface MlTrainingResult {
  success: boolean;
  message: string;
  models_trained: string[];
  training_samples: number;
  metrics?: MlModelMetric[];
}

export interface AssistantQueryInput {
  question: string;
  lat: number;
  lon: number;
  /** @nullable */
  city_name?: string | null;
}

export interface AssistantResponse {
  answer: string;
  confidence: string;
  weather_context: string;
  recommendations?: string[];
}

export interface HistoricalSummary {
  period: string;
  avg_temp: number;
  max_temp: number;
  min_temp: number;
  avg_humidity: number;
  total_rainfall: number;
  avg_wind_speed: number;
  avg_pressure: number;
  /** @nullable */
  avg_aqi?: number | null;
  daily_data: DailyWeather[];
}

export interface CityLocation {
  name: string;
  lat: number;
  lon: number;
  /** @nullable */
  country?: string | null;
}

export interface CityComparisonInput {
  cities: CityLocation[];
}

export interface CityComparisonResult {
  city: CityLocation;
  current_weather: CurrentWeather;
  air_quality: AirQuality;
  forecast_today?: DailyWeather;
}

export type GetCurrentWeatherParams = {
lat: number;
lon: number;
timezone?: string;
};

export type GetWeatherForecastParams = {
lat: number;
lon: number;
timezone?: string;
days?: number;
};

export type GetHistoricalWeatherParams = {
lat: number;
lon: number;
/**
 * YYYY-MM-DD
 */
start_date: string;
/**
 * YYYY-MM-DD
 */
end_date: string;
timezone?: string;
};

export type GetWeatherAlertsParams = {
lat: number;
lon: number;
};

export type GetAirQualityParams = {
lat: number;
lon: number;
};

export type SearchLocationsParams = {
q: string;
limit?: number;
};

export type ReverseGeocodeParams = {
lat: number;
lon: number;
};

export type GetMlPredictionsParams = {
lat: number;
lon: number;
};

export type GetMlModelsParams = {
lat: number;
lon: number;
};

export type GetHistoricalSummaryParams = {
lat: number;
lon: number;
period?: GetHistoricalSummaryPeriod;
};

export type GetHistoricalSummaryPeriod = typeof GetHistoricalSummaryPeriod[keyof typeof GetHistoricalSummaryPeriod];


export const GetHistoricalSummaryPeriod = {
  week: 'week',
  month: 'month',
  year: 'year',
} as const;