# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AERISYN – AI Weather Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cities/search-history:
    get:
      operationId: getSearchHistory
      tags: [cities]
      summary: Get recent search history
      responses:
        "200":
          description: Search history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchHistoryItem"
    post:
      operationId: addSearchHistory
      tags: [cities]
      summary: Add a city to search history
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchHistoryInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchHistoryItem"
    delete:
      operationId: clearSearchHistory
      tags: [cities]
      summary: Clear all search history
      responses:
        "204":
          description: Cleared
...
  /assistant/query:
    post:
      operationId: queryAssistant
      tags: [assistant]
      summary: Ask the AI weather assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantQueryInput"
      responses:
        "200":
          description: Assistant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantResponse"
...
components:
  schemas:
...
    SearchHistoryItem:
      type: object
      required: [id, name, lat, lon, country, searched_at]
      properties:
        id: { type: integer }
        name: { type: string }
        lat: { type: number }
        lon: { type: number }
        country: { type: string }
        country_code: { type: ["string", "null"] }
        admin1: { type: ["string", "null"] }
        timezone: { type: ["string", "null"] }
        searched_at: { type: string }
    SearchHistoryInput:
      type: object
      required: [name, lat, lon, country]
      properties:
        name: { type: string }
        lat: { type: number }
        lon: { type: number }
        country: { type: string }
        country_code: { type: ["string", "null"] }
        admin1: { type: ["string", "null"] }
        timezone: { type: ["string", "null"] }
...
    AssistantQueryInput:
      type: object
      required: [question, lat, lon]
      properties:
        question: { type: string }
        lat: { type: number }
        lon: { type: number }
        city_name: { type: ["string", "null"] }
    AssistantResponse:
      type: object
      required: [answer, confidence, weather_context]
      properties:
        answer: { type: string }
        confidence: { type: string }
        weather_context: { type: string }
        recommendations:
          type: array
          items: { type: string }
...